Move unit tests that aren't unit tests to a script
Catch2 is a unit test framework.
Unit tests are supposed to be to test specific features or functions of code, not the entire pipeline
We shouldn't expect user to run unit tests unless they are editing our code and want to make sure they haven't broke anything.
We currently have no unit tests.

Example should call the main executable with example data, I made a script to make this easier but we could also just add the inputs to the wiki